chore(invoices): remove dead VoidInvoiceDialog
Dialog was never triggered; void flow lives in CustomerInvoiceVoid page. Moved its gql mutation+fragment into that page.
